incident_env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "incident_env"
version = "1.0.0"
description = "Incident Response Triage Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[tool.setuptools.packages.find]
include = ["incident_env*"]
